Filter Expression parser
This commit implements a parser for $filter expressions, per the redfish
specification and odata specification.  This is intended to be used to
support $filter query for collections.

For parsing libraries, this commit chooses boost spirit x3.  It's chosen
because it doesn't require a new external dependency, and is done
entirely in the compiler, using C++ syntax.  While the syntax is still
somewhat difficult to read, there's a slew of unit tests included to
make sure that at least the common things we expect to work will parse
correctly.

Tested: Unit tests pass (good coverage).  Code not yet used.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I1b0ab615bc49064acab4dad47f0a8aa499557bfc