Use lexme in redfish filter parser
Previously, the parser added space ignore instructions between every
node.  This is because there was one place where we actually cared about
spaces, when doing operator comparisons (x eq y).  If spaces are
ignored, it's impossible to determine the end of x and the beginning of
eq.

Spirit x3 has a lexeme, which allows us to ignore the parser skips
temporarily, which allows us to parse the operations in a much simpler
way.  This also requires that we change to phrase_parse instead of
parse.

Tested: Unit tests pass.  Good coverage.

Change-Id: Ifc6f1681e8524ba5032ee118cc3b3a18b30c639e
Signed-off-by: Ed Tanous <ed@tanous.net>