Fix coredump when restart service if created subscriptions
When multiple subscriptions are created using the redfish url:
/redfish/v1/EventService/Subscriptions/, a coredump occurs during
the bmcweb restart process.

The problem was found when saving the subscription configuration
during the service stop process, remove the newly created shared
pointer 'subValue' and use the pointer in the map directly, this
problem was solved.

Tested: Create 18 subscriptions and restart the service 300 times
without coredump generation.

Change-Id: Ide90abdbc4c6e88215049783b88d9e91902c554d
Signed-off-by: wenlitao <w_litao@linux.alibaba.com>