Remove "Function wasn't a string?" messages for PCIe
Currently for every PCIe function bmcweb checks every property of the
"xyz.openbmc_project.Inventory.Item.PCIeDevice" interface if it is a
string. And if it is not the message "Function wasn't a string?" is
produced.
This is wrong since the properties "MaxLanes" and "LanesInUse" are not
strings but integers.
Drop the error print statement to remove the false error messages.

Change-Id: I1cc082e5aaf392b0cc4c051ab0bc6d8418aed0f9
Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>