Refactor HandleSubRoute to avoid req copy
The current Oem SubRoute is calling unconditionally for GET or PATCH.
As the result, if there is an error before handleSubRoute, the result
may still contain "Oem" data even if the operation may fail.

For example, for GET,  Oem/OpenBmc parts are still present.

```
% curl -k -X GET https://${bmc}/redfish/v1/Managers/bmc
{
  "Oem": {
    "OpenBmc": {
      "@odata.id": "/redfish/v1/Managers/bmc#/Oem/OpenBmc",
      "@odata.type": "#OpenBMCManager.v1_0_0.Manager",
      "Certificates": {
        "@odata.id": "/redfish/v1/Managers/bmc/Truststore/Certificates"
      }
    }
  },
  "error": {
    "@Message.ExtendedInfo": [
      {
        "@odata.type": "#Message.v1_1_1.Message",
        "Message": "The request failed due to an internal service error.  The service is still operational.",
        "MessageArgs": [],
        "MessageId": "Base.1.19.InternalError",
        "MessageSeverity": "Critical",
        "Resolution": "Resubmit the request.  If the problem persists, consider resetting the service."
      }
    ],
    "code": "Base.1.19.InternalError",
    "message": "The request failed due to an internal service error.  The service is still operational."
  }
}
```

This may be okay for GET, but PATCH may still cause the OemAction even
if the other part (e.g. valid dbus-path check) is failed.
For the proper handling OemSubRoute, it may need to occur inside
async-callback (which may check the valid dbus-path).

In this case, passing the original `req` to handleSubRoute
may need the copy of `req`. This can be avoid by creating `SubRoute`
structure before calling it.

Tested:
- Redfish Service Validator passes
- Manager GET/PATCH

Signed-off-by: Myung Bae <myungbae@us.ibm.com>